Refactor file structure and naming
